refactor: complete WebPageFactory split into 10 domain factories
WebPageFactory: 631 -> 136 lines (-78%)
9 new factory classes extracted, all zero- or single-dependency:

Zero-dependency: AuthPageFactory, ErrorPageFactory, SidebarFactory,
  SettingsPageFactory, SearchPageFactory, DevDashboardPageFactory

Service-dependency: ContactsPageFactory(contactService),
  HomePageFactory(messageService), InfraPageFactory(repository)

Existing: AdminPageFactory(securityService, notificationService)

All 20 public methods now delegate to domain factories via lazy vals.
Full backward compatibility - no consumer code changes needed.
